Rename main page to fix hot reloading in npm run dev
frontend/pages/index.tsx corresponds to the `/` route, which the broswer
kept failing to get
