+++ frontend/sections/Home.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Button, Heading, Text, VStack } from "@chakra-ui/react";
 
-import { Id } from "../pages/MainPage";
+import { Id } from "../pages";
 
 type Props = {
   id: Id;
