+++ frontend/sections/Newsletter.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Button, Heading, VStack } from "@chakra-ui/react";
 
-import { Id } from "../pages/MainPage";
+import { Id } from "../pages";
 
 type Props = {
   id: Id;
